Added a new file "convolution_tests.py". The idea is that MultiDimensionalLSTM should be implemented using 1D convolutation of the skewed input image in combination with 2D convolution using the same diagonal skewing of the input. It seems like in pytorch nn.Conv1D can do this. For each column of the image we would thus:
1. Compute a 1-D convolution of the image pixes (in fact this can also
be pre-computed for the entire image, for higher efficiency)
2. Compute a 2-D convolution of the two predecessor states.

The crucial part is that this will be done for all pixels in a column in
parallel, which is not possible when working with RNNCell. In other words,
we have to go to a lower level and implement the atomic operations of
MulitDimensional RNN in parallel for all pixels in one column, using
the parallelized convolution and other basic funtions.