Implemented MultDimensional RNN using convolution and image skewing, the tricks from the paper "Pixel Recurrent Neural Networks". This implementation should be much more efficient than an implementation using MDRNNCells, because with seperate RNN cells for each pixel, parallelization is a problem. By using full convolution layers as in the paper instead, in combination with the pixel skewing trick, the efficient CUDA methods can be used.
TODO: Right now things seem to work with batch size 1, but things don't
work yet for different batch sizes.
The example classifier from http://pytorch.org/tutorials/beginner/blitz/cifar10_tutorial.html
can work with batch size > 1. The idea would be that the computation needs to be done
for each of the images in the batch, and the results concatenated.
It is also a question if this can be further optimized:
rather than using for-loops, probably the computation can be done in parallel, using one
"channel" (i.e. the first dimension) per example.

	modified:   data_preprocessing/load_mnist.py
	modified:   modules/multi_dimensional_rnn.py
	modified:   util/image_input_transformer.py