Implemented a parallelized verions of the function creating skewed images, that does it for all the images in a batch at once.
Found out that rather than indexing tensors like this: [0][:][1]
the way to do this in pytorch is [0, :, 1] as the former does not
give the results you would expect, but the latter does.
See also:
https://stackoverflow.com/questions/47374172/how-to-select-index-over-two-dimension-in-pytorch?rq=1

Note: the function could even be further optimized, creating the skewed images
only once and saving them, then using the saved skewed images for every training epoch.

	modified:   modules/multi_dimensional_rnn.py
	modified:   util/image_input_transformer.py