Extracted a function "forward_multi_directional_multi_dimensional_function_fast" which can perform multi-directional computation for both MultiDimensionalRNN and MultiDimensionalLSTM using abstract methods
"def  _final_activation_function(self, final_activation_function_input)"

and

"def _compute_multi_dimensional_function_one_direction(self, function_input)"

which are implemented in the respective subclasses.
MultiDimensionalLSTM now works and has yielded already 91% accuracy on MNIST
on some runs, still without the multi-directional component.
The bias of all convolutions used in the forget gates is now set to 1, it
needs to be further investigated if this is indeed optimal.
See e.g. http://proceedings.mlr.press/v37/jozefowicz15.pdf
which suggests that a bias of 1 for the forget gate is advisable.

It is also still not really clear if the output activation function (tanh)
is nescessary or the output of the input and forget gates combined can
be directly merged with the output gate with out putting it through
the output activation function.