Refactored the code, removing the explicit wrapping in Variables and replacing it with an argument to the tensor constructors "requires_grad=True" or a call of the tensor function "tensor.requires_grad_(True)" Usage of the explicit Variable(tensor) method has become deprecated. More importantly, found the bug that made the computation fast but the results worse:
In the method "extract_unskewed_activations" in  modules/multi_dimensional_rnn.py
the line
"activations_unskewed = Variable(activations_unskewed)"
wrapped what was already a variable, again with a variable.
It remains somewhat unclear what exactly is the effect of this on the computation,
but conceptually it seems wrong to make a variable of what is already a variable,
and also it has a clear negative effect on the results of what is being learned.
Why exactly this is the case remains an open question.

After this fix the results are again better, but sadly running times are again much longer.

"
 # If activations_unskewed is made a variable (again!) it still works but runs
+        # much faster, but results are much worse somehow!!!
+        # print("activations_unskewed before: " + str(activations_unskewed.grad))
+        # print("activation_unskewed.requires_grad: " + str(activations_unskewed.requires_grad))
+        # activations_unskewed = Variable(activations_unskewed)
+        # print("activations_unskewed after: " + str(activations_unskewed.grad))
+
"

	modified:   modules/multi_dimensional_lstm.py
	modified:   modules/multi_dimensional_rnn.py
	modified:   modules/train_multi_dimensional_rnn.py