Added a python module for loading the CIFAR10 data. Adapted the multi_dimensional_lstm and image_input_transformer modules to accept multiple input channels, making them work for both MNIST and CIFAR10.
Accepting multiple inputs is important also when stacking layers of
MultiDimensionalLSTM  units: the first layer may have only one input channel as
input, but the following layers will have multiple channels,
hence support for multiple input channels is nescessary.

TODO: The next steps are:
Computing MultiDimensionalLSTM features block-wise over an
input tensor, with blocks of specified size:
	1. Splitting the input tensor into blocks
        2. Concatenating the blocks in the batch_size dimensions,
           so that computation using the blocks as input can be
           parallelized over blocks using the batch_size dimension
        3. Computing the MultiDimensionalLSTM features in parallel
           for each of the blocks
        4. Restructuring the outputs into the original input format,
           so that the outputs can be passed on to the next layer.

The idea is to next implement the simple MutliDimensionalLSTM block layer, followed
by dimensionality reduction layer architecture for handwriting recognition.
Finally, when a network is completed that computes a sequence of label probabilities,
still connectionist temporal classification needs to be added to complete things.
This adds some additional complication as it requires a special loss function,
but appearantly some ports of implementations of this for pytorch are available, so
this should be doable also.

	new file:   data_preprocessing/load_cifar_ten.py
	modified:   modules/multi_dimensional_lstm.py
	modified:   modules/multi_dimensional_rnn.py
	modified:   modules/train_multi_dimensional_rnn.py
	modified:   util/image_input_transformer.py