Create a new module "modules/block_multi_dimensional_lstm_layer_pair_stacking.py" for stacking multiple BlockMultiDimensionalLSTMLayerPair layers. This is to implement the main "body" of a handwriting recognition network. What still remains is in any case summing/merging the output columns and computing n-way softmax with continous temporal classification. It remains to be seen how to implement this exactly.
What is somewhat worrisome is that the performance of the new block-computation
networks falls behind with the earlier results of a single MultiDimensionLSTM
computed over the entire image inputs. It is not very clear why this is the
case (it happens for MNIST single digit recognition in particular), but
it should be investigated more.

	modified:   modules/block_multi_dimensional_lstm_layer_pair.py
	new file:   modules/block_multi_dimensional_lstm_layer_pair_stacking.py
	modified:   modules/block_strided_convolution.py
	modified:   modules/size_two_dimensional.py
	modified:   modules/train_multi_dimensional_rnn.py