Migrated the code for computing skewed/unskewed versions of tensors all to util/image_input_transformer.py, so that it is all in one place.
Added a method "forward_one_directional_multi_dimensional_lstm_with_additional_skewing_unskewing_step"
in MultiDimensionalLSTM, with the aim of testing whether the (old) methods that skew/unskew the
gradient using tensor slicing as opposed to "torch.cat" for patching together the skewed/unskewed
tensors cause problems to the gradient. However, the results of this are not conclusive.

Whereas the copying into a zeros tensor, produces "grad_fn" gradient fields that contain
a "CopySlices" object, rather than some "Backward" object, it is not clear whether or not this
is a problem. Possibly CopySlices is in fact also a backward object in disguise?

In either case, both methods seem to work at least for a single MultiDimensionalLSTM also
when the additional skewing/unskewing of the input is added, designed to stress-test the
skewing/unskewing methods for breaking the gradient. This suggests that breaking the gradient
may actually not be an issue, however, it is still good to use pytorch.cat, as it appears to
be faster than making a whole new tensor and copying over into that using tensor slicing.

	modified:   modules/multi_dimensional_lstm.py
	modified:   modules/multi_dimensional_rnn.py
	modified:   modules/train_multi_dimensional_rnn.py
	modified:   util/image_input_transformer.py