Added some intermediate test cases for networks with a single BlockMultiDimensionalLSTM plus BlockConvolutionLayer pair, followed by either: 1. A second BlockMultiDimensionalLSTM or 2. A second  BlockConvolutionLayer the results suggests that the second BlockMultiDimensionalLSTM is what makes training difficult, rather than the second BlockConvolutionLayer. This would make sense. However, increasing the number of epochs reveals that the setup with the second BlockMultiDimensionalLSTM also works quite ok, given enough training epochs:
"
...
[4,   200] loss: 0.854 Running time: 325.98467350006104
Number of gradient norm-based corrections: 0
[5,   100] loss: 0.768 Running time: 374.6045141220093
Number of gradient norm-based corrections: 0
[5,   200] loss: 0.720 Running time: 410.78008794784546
Number of gradient norm-based corrections: 0
[6,   100] loss: 0.647 Running time: 459.4248125553131
Number of gradient norm-based corrections: 0
[6,   200] loss: 0.612 Running time: 495.3232035636902
Number of gradient norm-based corrections: 0
[7,   100] loss: 0.562 Running time: 542.171270608902
Number of gradient norm-based corrections: 0
[7,   200] loss: 0.534 Running time: 577.0520710945129
Number of gradient norm-based corrections: 0
[8,   100] loss: 0.494 Running time: 623.9755685329437
Number of gradient norm-based corrections: 0
[8,   200] loss: 0.463 Running time: 658.8858370780945
Number of gradient norm-based corrections: 0
Finished Training
Accuracy of the network on the 10000 test images: 85 %

Process finished with exit code 0
"

What may be also a factor is not to increase the number of hidden
states for the second MultiDimensionalLSTM too much (?),
but the number of training epochs may be the main factor for success.
It might be the case that the two layer pair network actually also works
ok given more epochs, this is to be seen next.

	modified:   modules/block_multi_dimensional_lstm.py
	modified:   modules/block_multi_dimensional_lstm_layer_pair.py
	modified:   modules/block_multi_dimensional_lstm_layer_pair_stacking.py
	modified:   modules/block_strided_convolution.py
	modified:   modules/train_multi_dimensional_rnn.py