Added information to the readme about the repositories containing decoder code. There are two options basically 1. https://github.com/parlance/ctcdecode 2. https://github.com/SeanNaren/deepspeech.pytorch
The second repository subsumes the first repository, borrowing the beam search decoder
implemented in the first repository, as well as providing simpler greedy decoders etc.

Possibly the first repository provides already all that is needed, and the code from
the second repository won't be needed.

Another note: Using Adam as an optimizer seems to really have fixed the infinite loss
problems. When using no Xavier initialization it still works, but the loss starts
higher (so using the Xavier weight initialization for the linear layer is probably
still a good idea).

TODO: The next step now is to incorporate the decoder and after that write the
evaluation code, using the decoder.

	modified:   libraries/warp-ctc-installation-readme.txt