Implemented a first succesful approach to dealing with variable length sequences. Learning works, and after 2 epochs of training of sequences of randomly length 1 or 2, the programs obtains 11% accuracy:
"Accuracy of the network on the 10000 test images: 11 %

Process finished with exit code 0
"

The current implementation adds padding to the examples in
load_mnist, and uses negative values of -2 for the padding
of the labels. The actual label sequence length is then also
used to infer the actual length of the probabilities output
(the part of the sequence that is not padding). However, this
requires a fixed relation between the two, e.g. having always
4 rows in the probability matrix per digit. This is a problematic
assumption for real examples.
The solution seems to be in padding the images with negative values
also rather than with zeros, so that the sizes for the probabilities
output can be directly inferred from those. This would go in combination
with usage of "padd_packed_sequence", see:
https://discuss.pytorch.org/t/solved-training-lstm-by-using-samples-of-various-sequence-length/9641
https://pytorch.org/docs/master/_modules/torch/nn/utils/rnn.html
https://discuss.pytorch.org/t/understanding-pack-padded-sequence-and-pad-packed-sequence/4099/7

The latter causes the computation to actually be avoided for the padding part of the input, which
is what you want. This shoudl in principle allow using any values, including negative values
for the paddin, without any consequence. It might also make the computations faster.

	modified:   ctc_loss/warp_ctc_loss_interface.py
	modified:   data_preprocessing/load_mnist.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py