Changed the way the information about the labels_sequence_length and probabilities_sequence_length is passed to the training loop and to the evaluation loop. This information is now added add the end of the labels sequence inside the load_mnist code. The two numbers are stored as negative numbers furthermore, so that they will not be confused with actual labels. There is still some trickery nescessary, since what is needed is the size of the real output probabilities produced by the network, but what can be stored only is the size of the real input (i.e. its width). For this reason, a "width_reduction_factor" must be defined and used to compute to what actual output width the original real input width will be reduced after processing by the network.
While still not perfect, the changes do help to get rid of a lot of
obscure code to reconstruct real sizes based on the amount of -1
elements etc. Simply storing these sizes at the end of the labels
list is a more elegant solution. (The alternative would be
reimplementing data loader, so that it can output also
a tensor with the real data sizes in addition to the data and labels,
However, reimplementing data loader is quite involved, so the above
is a simpler solution that also works good enough.)

	modified:   ctc_loss/warp_ctc_loss_interface.py
	modified:   data_preprocessing/load_mnist.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py