Added timing information to the ctc loss training. It becomes clear that the training takes really long on the eam dataset:
"...
Time used for network forward: 0m 2s
Time used for ctc loss computation: 0m 0s
loss: tensor([ 0.])
Time used for loss backward: 0m 36s
Time used for this batch: 0m 44s
Processed 96 of 9074 examples in this epoch
>>> Total time used during this epoch: 2m 15s (- "expected time remaining: 212m 12s)
...
"
(This is with two GPUs and a batch size of 32. With only one GPU, batch size 16 only works because
of memory limitations. Essentially when running on multiple GPUs the batch is split across GPUs
and each GPU processes its assigned part of the batch.)

In contrast, in the paper "Handwriting Recognition with Large Multidimensional Long Short-Term
Memory Recurrent Neural Networks" (Voigtlander et.al, 2016) manages to do 1 epoch in 15 minutes
on a single GPU. So that is still approximately 28 times faster.

TODO: One observation is that currently the number of features is multiplied by a factor of 4
for every consecutive MDLSTM layer (as a function of the first layer). However, in the
(Voigtlander. et al, 2016) work, they use HiddenUnits * n, with n the layer number.
So in this case the 2nd MDLSTM layer would get 3 * HiddenUnits , and the 3th layer
5 * HiddenUnits instead of 4 * HiddenUnits and 4 * 4 * HiddenUnits = 16 * HiddenUnits.
So changing this in my implementation should increase speed by at least a factor (16 / 5 = 3.2)
Also, the paper writes that "we vary the number of hidden units per layer from 5n to 30n" and
"... we decided to limit the number of hidden units per layer to a maximum of 120".
Hopefully, decreasing the hidden units size will also increase the supported batch size, further
speeding things up.