Fixed the problem with zero loss in  modules/train_multi_dimensional_rnn_ctc.py. The problem occured because a wrong value for the "horizontal_reduction_factor" (now renamed to "width_reduction_factor") was  passed to WarpCTCLoss.compute_ctc_loss. As a consequence, a wrong value for the "real" length of the probabilities was computed by the function WarpCTCLoss.get_real_probabilities_length  . This error in turn happened, because now with a network with three layer pairs (pairs of
BlockMDLSTM + BlockStridedConvolution), with block sizes for the BlockStridedConvolution of
(4,2) the reduction factor is 2 * 2 * 2 instead of 2 * 2 (for 2 layer pairs).
The error has been fixed now in a more robust way, by adding a new function "get_width_reduction_factor"
to "BlockMultiDimensionalLSTMLayerPairStacking" which properly computes the reduction factor, and
querying this function.

Additionally, a check has been added to the function WarpCTCLossInterface.get_real_probabilities_length,
which checks that the computed "real" probabilities length is not bigger than the probabilities_tensor
sequence length, which should never happen, and is indicative of an erroneous "width_reduction_factor".

Note that the computation of these real length using the "width_reduction_factor" is not as pretty,
as one would like. There is however not really another good way to do it, since the data loader /
data set simply does not know yet what network the data will be used for during creation,
and the downstream network, by reducing the width, will determine the
output "real" lengths, which are what matters for the WarpCTCLossInterface.compute_ctc_loss function.

	modified:   ctc_loss/warp_ctc_loss_interface.py
	modified:   modules/block_multi_dimensional_lstm_layer_pair_stacking.py
	modified:   modules/block_strided_convolution.py
	modified:   modules/multi_dimensional_rnn.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   util/tensor_utils.py