Updated the to-do-list. Basically, it seems that training on the IAM dataset now works, but it is still very slow (in particular the backward function).
TODO:
Decreasing the number of hidden units in downstream layers did not
really seem to affect this either. But finding a way to speed up
training performance is now crucial for getting a workable system.

This may be tough, since there is no direct way to optimize the
backward function.
One idea is to test a version of the network without the ctc loss
function, but a multi-class output (as in MNIST single digit
classification) instead. Then see what is the speed of the different
steps in the computation. This might help in determining whether use
of the warp_ctc_loss has a large role in the slowness of the backward
function, or this is not so much caused by that but simply by the
(main) network structure and the size of the input.

	modified:   ctc_loss/warp_ctc_loss_interface.py
	modified:   todo-list.txt