Changed the IAM dataset loader so that it stores the images in the original unsigned int (uint8) format. The transformation to floats is then done on the gpu, after moving the original data in uint8 format to the GPU. This is not so much faster, rather more or less as fast, but it does save quite some working memory. The reason is that now the dataloader saves everything in memory in uint8 format, which has a smaller memory footprint, and the expansion to float is only done for the batch that is being processed.
Extracted the separate class file "network_to_sogt_max_network" and improved the
code with a new strategy class "ActivationsResizer" which helps to merge
multiple activations rows by either:
1. Summation (old recipe)
2. Keeping all rows, concatenation on channels dimensions

The new strategy of concatenation on channels dimension seems to give better
preleminary results, in the sense of faster decreasing loss. This would make
sense, since it is expected that by summation of the activations rows, a
lot of information may be lost.

	modified:   data_preprocessing/iam_database_preprocessing/iam_lines_dataset.py
	modified:   modules/block_multi_dimensional_lstm_layer_pair_stacking.py
	modified:   modules/block_strided_convolution.py
	modified:   modules/multi_dimensional_rnn.py
	new file:   modules/network_to_softmax_network.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py