Added an implementation of model-internal clipping / clamping of the gradients, using hooks. This is inspired on the remark that: "“Note that clip_grad_norm_ modifies the gradient after the entire backpropagation has taken place. In the RNN context it is common to restrict the gradient that is being backpropagated during the calculation." - Thomas Viehmann
And adapted from examples of:
https://discuss.pytorch.org/t/proper-way-to-do-gradient-clipping/191/15
https://github.com/t-vi/pytorch-tvmisc/blob/master/misc/graves_handwriting_generation.ipynb
https://discuss.pytorch.org/t/register-backward-hook-on-nn-sequential/472
https://github.com/DingKe/pytorch_workplace/blob/master/rnn/modules.py#L122

and Alex Graves paper, which mentions this method very briefly:
 https://arxiv.org/abs/1308.0850

The idea is that the gradient of tensors in the MDLSTM blocks mainly
is clipped just before the sigmoid and tanh activation functions.
This works in as much as that infinite loss can be avoided with it
without using gradient norm clipping (over the full gradient, i.e.
after calling loss backward).

However, the clipping/clamping range must be really narrow in order to
avoid the infinite losses from warp_ctc loss (i.e. to avoid divergence).
This works for example with CLAMPING_BOUND = 0.01.
But normally bounds of -10, 10 or -100, 100 for this kind of gradient
clamping are more common, so it seems that something is still not really
OK.

Finally changed the MultiDimensionalLSTM to actually use the output gate,
which was not done properly before.

	modified:   modules/block_multi_dimensional_lstm.py
	modified:   modules/block_multi_dimensional_lstm_layer_pair.py
	modified:   modules/block_multi_dimensional_lstm_layer_pair_stacking.py
	modified:   modules/block_strided_convolution.py
	new file:   modules/inside_model_gradient_clipping.py
	modified:   modules/multi_dimensional_lstm.py
	modified:   modules/multi_dimensional_lstm_parameters.py
	modified:   modules/multi_dimensional_rnn.py
	modified:   modules/network_to_softmax_network.py
	modified:   modules/parallel_multiple_state_weightings_computation.py
	modified:   modules/train_multi_dimensional_rnn.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py