Improved the output. Removed superfluous output from the MNIST data loader, added reporting of the average gradient total norm to the training script. It appears that the average gradient total norm is increasing somewhat during training on the variable length MNIST digit sequences, but possibly it also seems to saturate after a while, with a value around 20000.
"
Average gradient total norm: tensor(16561.4551, device='cuda:0')
Processed 28160 of 29994 examples in this epoch
>>> Time used in current epoch: 1m 1s (- "expected time remaining: 0m 4s)
[9,   230] loss: 3.469 Running time: 631.6995573043823
Number of gradient norm-based corrections: 10
Average gradient total norm: tensor(16122.7422, device='cuda:0')  <====
Processed 29440 of 29994 examples in this epoch
>>> Time used in current epoch: 1m 4s (- "expected time remaining: 0m 1s)
"

In any case, the xavier weight initialization throughout appears to
be working.
Another conclusion is that at least on the MNIST dataset, using
a single-directional LSTM seems to be not a problem, the network is
still learning. Starting from 0% accuracy it reaches 35% accuracy in
20 epochs:

i.e. for random sequences of length 0-3
"
Accuracy of the network on the 10000 test inputs: 27 %
</evaluation epoch 9 >
[11,    10] loss: 3.393 Running time: 710.9828922748566
...

Accuracy of the network on the 10000 test inputs: 29 %
</evaluation epoch 12 >
...
Accuracy of the network on the 10000 test inputs: 30 %
</evaluation epoch 13 >
...
Accuracy of the network on the 10000 test inputs: 35 %
</evaluation epoch 19 >
"
This suggests that using the single-directional network on the
IAMT network (which is faster), should not be the reason that
saturation of the loss is being observed.

TODO: One difference between the MNIST network and teh IAMT network
is the latter has less parameters. If the problems with
loss saturation persist this might be something else to look at.

	modified:   data_preprocessing/load_mnist.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py