Refactored the code, adding a Trainer class to take care of training one epoch, and also a separate class for gradient clipping. This refactoring is also in preparation of properly saving and loading the model, optimizer etc, so that training can resume from an earlier checkpoint.
Also found and fixed a bug in the application of the norm-based gradient clipping:
the clipping was done not on the entire network, but only on the network excluding
the last linear layer (that prepares the input for warp_ctc).
This uneven clipping could be expected to mess up the results somewhat, and with this
fixed the results could still get better.

	modified:   data_preprocessing/iam_database_preprocessing/iam_lines_dataset.py
	new file:   modules/gradient_clipping.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	new file:   modules/trainer.py