Further refactored the code, adding a class for evaluation. Also changed the method "   def get_random_train_set__test_set_data_loaders" " in  data_preprocessing/iam_database_preprocessing/iam_lines_dataset.py to " def get_random_train_set_validation_set_test_set_data_loaders".
The new method returns a train loader, validation loader and test loader
(fractions of train data, test data and test data, which must add up to 1,
are provided to the method as arguments).

Changed the code to use this validation data during the training loop for
validation.

TODO: The next step is to drop checkpoints. The validation scores will be
useful in this.

	new file:   config/__init__.py
	modified:   data_preprocessing/iam_database_preprocessing/iam_lines_dataset.py
	new file:   modules/evaluator.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   modules/trainer.py