Completed the work on creating and using an iam words dataset, as an alternative to the lines dataset. Testing on the words dataset should be faster, and training should be easier. As such this is an intermediate test case between MNIST random sequence recognition and fullfledged line recognition, which is still showing problems.
Added also functionality for optionally enforcing minimum height/width dimensions,
in order to filter examples that are probably erroneous. However, for the words
database, some bounding boxes for punctuation can be legitimately very small.
To avoid bugs with scaling the images down, added a check to see if the size
of the to be scaled image is at least as big as the scale reduction factor.
If not, scaling is skipped for that image. (After scaling padding is still
done, so in either case the padded images will be the same size.)

	renamed:    data_preprocessing/iam_database_preprocessing/iam_lines_dataset.py -> data_preprocessing/iam_database_preprocessing/iam_dataset.py
	new file:   data_preprocessing/iam_database_preprocessing/iam_examples_dictionary.py
	deleted:    data_preprocessing/iam_database_preprocessing/iam_lines_dictionary.py
	modified:   modules/evaluator.py
	modified:   modules/opts.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   modules/trainer.py