Changed the method that computes the optimized scale reduction factor, so it does not increase the scale reduction factor over the minimum specified, if by doing so it cannot save more than half an "output row" (corresponding to input rows consumed to produce the output row). This avoids that the input gets scaled down excessively just to get minor savings in reducing vertical padding.
More importantly made a bug fix of the method "get_real_probabilities_length"
in "warp_ctc_loss_interface".
This method should round the result of real_width / horizontal_reduction_factor
up to the smallest larger or equal integer. This was not done, instead the
result was just round down by implicitly casting to integer. That is  problematic,
since it means that valid output would be lost whenever the real input length
would not be an exact multiple of horizontal_reduction_factor (currently 8),
meaning the network output of up till 7 pixel columns of actual input coul be lost.

	modified:   ctc_loss/warp_ctc_loss_interface.py
	modified:   data_preprocessing/iam_database_preprocessing/iam_dataset.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py