Fixed a severe bug in the code relating to the way string_to_index_mapping_table was filled. The static method "def get_examples_line_information" in data_preprocessing/iam_database_preprocessing/iam_dataset.py would get an enumeration over the line dictionary elements. But because these dictionaries were normal and not ordered dictionaries, the order of the lines would be non-deterministic and changing every time. As a consequence, the lines would be provided to the string_to_index_mapping_table in an every time changing order, leading to an every time different table. This would cause problems with loading from a checkpoint, as the string_to_index_table would be changed from the original used when saving the checkpoint, and hence the results would be severely compromised.
The fix used now is to use OrderedDict instead of dic in
"data_preprocessing/iam_database_preprocessing/iam_examples_dictionary.py".
This fix has been tested to work, however it is still brittle, as it stil
requires the order of the lines in the original training set to remain
exactly the same.

TODO: A safer option would probably be to save the string_to_index_mapping_table to a file,
just as he permutation is saved. Recovering the string_to_index_mapping_table
can then be done using the same file, when loading from a checkpoint, which
will assure the order is preserved even if the order of the training examples would be
changed for example.

	modified:   data_preprocessing/iam_database_preprocessing/data_permutation.py
	modified:   data_preprocessing/iam_database_preprocessing/iam_dataset.py
	modified:   data_preprocessing/iam_database_preprocessing/iam_examples_dictionary.py
	modified:   data_preprocessing/iam_database_preprocessing/string_to_index_mapping_table.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   modules/trainer.py