Started implementing the possibility to use minimal horizontal padding for examples of different sizes, by padding only to a uniform size in the vertical dimension and for each example seperate to a multiple of the network width_reduction_factor in the horizontal dimension.
On the data preparation side,
this required several changes to the  data_preprocessing/iam_database_preprocessing/iam_dataset.py
class, including the implementation of PaddingStrategy classes, that implement the padding
of the dataloader. Additionally a custom collate function needs to be implemented and passed to
the dataloader, see: https://discuss.pytorch.org/t/how-to-create-a-dataloader-with-variable-size-input/8278/3
for a discussion on this point.

On the network computation side, this required a new version of tensor chunking called tensor list chunking,
which chunks lists of tensors of varying size into one tensor of blocks concatenated along a batch dimension,
and another function to reconstruct the original list. In the case of the block-strided convolution, the output
size actually changes (not only the channel dimension), so I also implemented a possibility to specify a different
output (block) size.

The class util/tensor_list_chunking.py is crucial for the new computation without padding, and has been tested
with new test code, to see if the functions manage to chunk and reconstruct the original input from the chunking.

TODO: Things seem to work up to a point, but the class modules/network_to_softmax_network.py
still has to be adapted to deal with the new input as a list. If the input is a list, the forward function
of this module should compute the activations, then concatenate them using block chunking, then compute the linear activations,
and finally convert back. This is very similar to what is done in block_strided_convolution.

	modified:   data_preprocessing/iam_database_preprocessing/iam_dataset.py
	modified:   data_preprocessing/iam_database_preprocessing/iam_examples_dictionary.py
	modified:   modules/block_multi_dimensional_lstm.py
	modified:   modules/block_multi_dimensional_lstm_layer_pair.py
	modified:   modules/block_multi_dimensional_lstm_layer_pair_stacking.py
	modified:   modules/block_strided_convolution.py
	modified:   modules/network_to_softmax_network.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   modules/trainer.py
	new file:   util/tensor_list_chunking.py