Made remaining changes to be able to train and test with variable length images without padding everything to the same width. This required changing the forward functions of various modules to have adapted code for the case that the input is a list. Adapted the MNIST data loaders so that they can also be used with this new functionality. Things seem to be working somewhat but there are still a couple of problems:
1. It is still generally slow. Both the block_multi_dimensional_lstm and
block_strided_convolution get input and produce outputs as list, and
compute internally with a single tensor by using tensor_list_chunking.
But the packing into the single tensor before computation, and unpacking after
the computation is probably expensive, and has not been optimized yet.
Right now, for MNIST, on a single gpu, the old version with full padding is
still about 2 times faster than the version with minimal horizontal padding.

2. The computation with minimal horizontal padding is not well supported by
DataParallel, since it cannot slice up the input lists and divide the labor
over several GPUs: it can only do this if the input is a tensor.
This could probably be fixed by making a custom version of DataParallel, with
alternative scatter and gather functions, see this discusssion:
https://discuss.pytorch.org/t/dataparallel-chunking-for-a-list-of-3d-tensors/15962.
In the current use case, the original input is a list - that should be sliced,
but the final output is a tensor. To deal with this case correctly will require
wuite some work.

TODO:
1. See if the tensor_list_chunking can be optimized for speed.
2. Change the code to avoid unnesscessary de-chunking and re-chunking steps
   between block_multi_dimensional_lstm and block_strided_convolution_layers.
   To be able to do this, in addition to the input, a list with lengths should
   also be passed to be able to produce again the right-size outputs from the
   block_strided_convolution (when a list is received as input, this information
   is present in the list, but to avoid additional de-chunking + re-chunking,
   block_strided_convolution would receive the chunked output from
   block_multi_dimensional_lstm.

3. See if a custom DataParallel can be made that parallelizes properly in case
   of list rather than tensor input to the forward function.