Slightly increased the performance of the list chunking function.
TODO:
But the big savings seem to be in just avoiding unnesscessary
chunking and dechunking as much as possible. A obervation is that
the layout of the blocks does not matter for the results, as long
as the output blocks of the final layer are put back to correspond
to the original input order. This means it would be smarter to
avoid the dechunking and chunking altogether within the main network
layers, performing the chunking just before and dechunking just
after the main computation and before the linear layer (for softmax)
computation. This would reduce the chunking+dechunking operations
from 6 to one for a 3 layer network, and from 4 to 1 for a 2 layer
network, which should yield considerable savings since these computations
are expensive.

	modified:   modules/block_multi_dimensional_lstm.py
	modified:   modules/block_strided_convolution.py
	modified:   util/tensor_list_chunking.py