Added functionality to deal with non-variable number of output rows when minimizing horizontal padding.
Fixed a big bug in the dechunking code in
util/tensor_list_chunking.py
in the method "dechunk_block_tensor_concatenated_along_batch_dimension_changed_block_size":

"
 for row_index in range(1, blocks_per_column):
-                tensor_block_row = TensorListChunking.reconstruct_tensor_block_row(row_slices[0])
+                tensor_block_row = TensorListChunking.reconstruct_tensor_block_row(row_slices[row_index])
                 reconstructed_example_tensor = torch.cat((reconstructed_example_tensor, tensor_block_row), 1)
"

in the for loop code, used the wrong index for the call to "TensorListChunking.reconstruct_tensor_block_row",
namely 0 instead of row_index.
Because of this, the same block row would be concatenated multiple times to the reconstructed output, messing
upt the results in case there are two ouptut rows (which previously, with full vertical padding, was always
the case). Added debugging code that helped finding this bug, checking if tensors are equal or not and visualizing
these unexpected equal or not equal tensors as images.

With this fix, the code seems to run with the minimal vertical padding, and even seems to be giving more varied
output than before. Thus it appears that this may be another step in the direction of fixing the flaw were
the output only learns to output "e" - having less vertical padding should make the training material more
informative and aid in more effective learning.

	modified:   modules/network_to_softmax_network.py
	modified:   util/image_visualization.py
	modified:   util/tensor_list_chunking.py
	modified:   util/tensor_utils.py