Added the functionality to have {multi-dimensional lstm, block-strided convolution} layer pairs, as opposed to {block-multi-dimensional lstm, block-strided conbolution} layer pairs. As it turns out, it is not at all a good idea to compute the mdlstm features only restricted to blocks, to benefit from mdlstm properly it needs to be computed over the entire image or entire input map for each layer. This appears to be working now ok, and already the network is producing more interesting output early on
2 epochs on small data:
">>> evaluate_mdrnn  - output: ""
reference: "A" --- wrong
>>> evaluate_mdrnn  - output: ""
reference: "." --- wrong
>>> evaluate_mdrnn  - output: "eeoeoeoeD"
reference: "Nkumbula" --- wrong
>>> evaluate_mdrnn  - output: "eee"
reference: "West" --- wrong
>>> evaluate_mdrnn  - output: "e"
reference: "his" --- wrong
>>> evaluate_mdrnn  - output: "eeoeo"
"
3 epochs on small data:
"
>>> evaluate_mdrnn  - output: "o"
reference: "they" --- wrong
>>> evaluate_mdrnn  - output: "orrrar"
reference: "America's" --- wrong
>>> evaluate_mdrnn  - output: "orrrarar"
reference: "association" --- wrong
>>> evaluate_mdrnn  - output: "o"
reference: "the" --- wrong
>>> evaluate_mdrnn  - output: "orrro"
reference: "Macleod" --- wrong
>>> evaluate_mdrnn  - output: "or"
reference: "And" --- wrong
"

Still not Shakespeare, but definetely
much better than only "e" ...

One concern is that the total gradient norm becomes nan,
"Average gradient total norm: tensor(nan., device='cuda:0')"
Perhaps this could be due to some of the gradients becoming zero
because of the long length of the sequence (vanishing gradient problem?)

It remains to be seen if this problem also persists with more data
(probably it would, as it appears right from the start in training)
and if this is a problem that really hinders learning (it might very well be).

TODO: Additionally, found a problem in the implementation of MDLSTM. The
computation should use a mask to keep track of which input cells and more importantly which
predecessor cells are valid. In the skewed input configuration, used for fast computation
with convolution, many of the cells are not valid and should be ignored when queried
as predecessor hidden states or memory states. However, this has not been implemented.
What needs to be done is to create a mask tensor, with zeros for invalid cells and ones
for valid cells. This mask tensor can then be queried to multiply the adequate variables
with during MDLSTM forward computation, to remove elements from the computation that correspond
to invalid predecessor states. This should hopefully again further increase the effectiveness
of learning.

	modified:   data_preprocessing/iam_database_preprocessing/iam_dataset.py
	modified:   data_preprocessing/iam_database_preprocessing/string_to_index_mapping_table.py
	deleted:    modules/block_multi_dimensional_lstm_layer_pair.py
	new file:   modules/mdlstm_layer_block_strided_convolution_layer_pair.py
	modified:   modules/multi_dimensional_lstm.py
	renamed:    modules/block_multi_dimensional_lstm_layer_pair_stacking.py -> modules/multi_dimensional_lstm_layer_pair_stacking.py
	modified:   modules/network_to_softmax_network.py
	modified:   modules/train_multi_dimensional_rnn.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   modules/trainer.py